Modifies default rails generate scaffold files
Rails generate commands can be massively helpful. They can also give
us more than we need. In this event, we're not using "Rails Views"
so we don't need the 'new' or 'edit' routes. We're also not using
`location` so we don't need that in the `create` method. We also care
that our data is uniform, so we don't want any doctors to be created
that don't have at a minimum first and last names.